Allow to verify liquidity provision tx hashes before executing (#219)
Closes #193

This PR makes it so that we can verify the tx hash of a liquidity provisioning transaction before signing it in the safe interface.

This is achieved by adding an optional `--dry-run` arguments which instead of sending out transactions only simulates them. In combination with the `--bracket ` argument introduced from #210 this allows reproducing the exact transaction hash that the original proposer created.

I also added another optional `--nonce` argument to allow verifying hashes even when the safe has moved on signing other txs and thus gotten a new nonce. In the regular case where we verify the tx before signing this shouldn't be needed.

Note, that this verification method is still not 100% fool proof, since the web interface doesn't seem to show the tx hash anywhere. Therefore we still trust that the proposer is sharing the hash with use that corresponds to the calldata that was submitted in the interface.

### Test Plan

Run the regular liquidity provision command:

```
origin/191/use_existing_bracketsnpx truffle exec scripts/complete_liquidity_provision.js --targetToken=0 --stableToken=7 --lowestLimit=0.10 --highestLimit=1 --currentPrice=0.70 --masterSafe=$MASTER_SAFE --investmentTargetToken=0.05 --investmentStableToken=0.05 --fleetSize=20 --network=$NETWORK_NAME
```

Then re-run is ammended with the arguments provided as a last statement of the first command (`--nonce` is not required).

See that the tx hashes are the same in both commands.