Use GnosisSafe npm artefact instead of local compilation (#528)
As discussed on #518 relying on our local compilation of dependent contracts can be error-prone if we e.g. update the GnosisSafe contract but don't remember to do a full local "re-migration".

This PR is a first step to removing our external dependecies (Safe, SafeFactory, MultiSend, BatchExchange) by using the artefact from the npm package directly.

The challenge with this was that in tests, we need a locally deployed instance (so that we can call GnosisSafe.deployed even on ganache). To achieve this we now use artefactor to write the locally deployed json back into the dependent npm folder.

Also tagging @rmeissner as he might have some idea of how to better accomplish what we want.

### Test Plan

complete liquidity provision still works on Rinkeby + unit tests pass